Move ServiceAccount handling to own file
Since the SA code will be common between rsync & rclone, it shouldn't be
in rsync_common.go
